third_party/WebKit/LayoutTests/vr/requestPresent_resolve_webgl2.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl2-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  if (!gl) {
    // WebGL 2 is not supported. This is legal.
    t.done();
    return Promise.resolve();
  }

  return navigator.getVRDisplays().then( (displays) => {
    var display = displays[0];

    runWithUserGesture( () => {
      display.requestPresent([{ source : webglCanvas }]).then( () => {
        t.step( () => {
          assert_true(display.isPresenting);
        }, "Display should be presenting");
      }, (err) => {
        t.step( () => {
          assert_unreached(err);
        }, "Should never reach here");
      }).then( () => {
        t.done();
      });
    });
  }, (err) => {
    t.step( () => {
      assert_unreached(err);
    }, "getVRDisplays rejected");
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"Test requestPresent resolves when provided a WebGL2 canvas");

</script>
